initial trend modularization
